Use OSC Helm charts for release/snapshot installation
it/dep helm charts are published to OSC Helm repositories.
SNAPSHOT repository is available with the charts and used.
Once the release charts are available. This needs to be modified to support that.

Issue-ID: NONRTRIC-1076

+++ smo-install/scripts/layer-0/0-setup-helm3.sh
@@ -6,6 +6,7 @@
 # ================================================================================
 # Copyright (C) 2021 AT&T Intellectual Property. All rights
 #                             reserved.
+# Copyright (C) 2025 OpenInfra Foundation Europe. All rights reserved.
 # ================================================================================
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -23,25 +24,66 @@
 #
 ###
 #Helm package
-cd /tmp
-wget https://get.helm.sh/helm-v3.12.3-linux-amd64.tar.gz
-tar xvfz /tmp/helm-v3.12.3-linux-amd64.tar.gz
-sudo mv linux-amd64/helm /usr/local/bin/helm
-sudo apt-get install git -y
+SCRIPT=$(readlink -f "$0")
+SCRIPT_PATH=$(dirname "$SCRIPT")
 
+# Check whether helm is installed and if not install it
+if command -v helm > /dev/null 2>&1; then
+    echo "Helm is already installed. Skipping installation."
+else
+    echo "Helm is not installed. Installing helm ..."
+    cd /tmp
+    wget https://get.helm.sh/helm-v3.12.3-linux-amd64.tar.gz
+    tar xvfz /tmp/helm-v3.12.3-linux-amd64.tar.gz
+    sudo mv linux-amd64/helm /usr/local/bin/helm
+fi
 
 echo "Checking HELM ..."
 helm version
 
-TAR_VERSION=v0.10.3
-echo "Downloading and installing helm-push ${TAR_VERSION} ..."
-TAR_FILE=helm-push-${TAR_VERSION}.tar.gz
-HELM_PLUGINS=$(helm env HELM_PLUGINS)
-mkdir -p $HELM_PLUGINS/helm-push
-cd $HELM_PLUGINS/helm-push
-wget https://nexus.o-ran-sc.org/content/repositories/thirdparty/chartmuseum/helm-push/$TAR_VERSION/$TAR_FILE
-tar zxvf $TAR_FILE >/dev/null
-rm $TAR_FILE
-cd /tmp/
-helm repo remove local
-helm repo add local http://localhost:18080
+# Check whether git is installed and if not install it
+if command -v git > /dev/null 2>&1; then
+    echo "Git is already installed. Skipping installation."
+else
+    echo "Git is not installed. Installing git ..."
+    sudo apt-get install git -y
+fi
+
+# Check whether helm-push is installed and if not install it
+if helm plugin list | awk '{print $1}' | grep -w -q "cm-push"; then
+    echo "Helm cm-push plugin is already installed. Skipping installation."
+else
+    echo "Helm cm-push plugin is not installed. Installing cm-push ..."
+    TAR_VERSION=v0.10.3
+    echo "Downloading and installing helm-push ${TAR_VERSION} ..."
+    TAR_FILE=helm-push-${TAR_VERSION}.tar.gz
+    HELM_PLUGINS=$(helm env HELM_PLUGINS)
+    mkdir -p $HELM_PLUGINS/helm-push
+    cd $HELM_PLUGINS/helm-push
+    wget https://nexus.o-ran-sc.org/content/repositories/thirdparty/chartmuseum/helm-push/$TAR_VERSION/$TAR_FILE
+    tar zxvf $TAR_FILE >/dev/null
+    rm $TAR_FILE
+fi
+
+# Check whether make is installed and if not install it
+if command -v make > /dev/null 2>&1; then
+    echo "Make is already installed. Skipping installation."
+else
+    echo "Make is not installed. Installing make ..."
+    sudo apt-get install make -y
+fi
+
+cd $SCRIPT_PATH
+# Check whether undeploy and deploy plugins are installed by comparing exact name and if not install them
+if helm plugin list | awk '{print $1}' | grep -w -q "undeploy"; then
+    echo "Helm undeploy plugin is already installed. Skipping installation."
+else
+    echo "Helm undeploy plugin is not installed. Installing undeploy plugin ..."
+    helm plugin install ../../onap_oom/kubernetes/helm/plugins/undeploy/
+fi
+if helm plugin list | awk '{print $1}' | grep -w -q "deploy"; then
+    echo "Helm deploy plugin is already installed. Skipping installation."
+else
+    echo "Helm deploy plugin is not installed. Installing deploy plugin ..."
+    helm plugin install ../../onap_oom/kubernetes/helm/plugins/deploy/
+fi
